feat(tasks): Extracts changing task status to separate module. Validates new task on response creation
