+++ taskforce/apps/tasks/src/app/response/response.const.ts
@@ -1,1 +1,3 @@
-export const RESPONSE_ALREADY_EXISTS = 'Contractor reponse on this task already exists';
+export const RESPONSE_ALREADY_EXISTS =
+  'Contractor reponse on this task already exists';
+export const NOT_NEW_TASK = 'Can respond only to new tasks';
