+++ taskforce/apps/tasks/src/app/response/response.module.ts
@@ -2,8 +2,10 @@ import { Module } from '@nestjs/common';
 import { ResponseService } from './response.service';
 import { ResponseController } from './response.controller';
 import { ResponseRepository } from './repository/response.repository';
+import { TaskModule } from '../task/task.module';
 
 @Module({
+  imports: [TaskModule],
   controllers: [ResponseController],
   providers: [ResponseRepository, ResponseService],
   exports: [ResponseService],
