+++ taskforce/apps/tasks/src/app/task-status/task-status.const.ts
@@ -1,0 +1,5 @@
+export const INVALID_COMMAND_ERROR =
+  'command cannot be applied to current task status';
+export const NO_RESPONSE_CONTRACTOR_ERROR =
+  'Contractor has not responded to task';
+export const CONTRACTOR_NOT_FREE_ERROR = 'Contractor has another task to do';
