+++ taskforce/apps/tasks/src/app/task/task.const.ts
@@ -1,9 +1,4 @@
 export const TASK_NOT_FOUND_ERROR = 'task not found';
-export const INVALID_COMMAND_ERROR =
-  'command cannot be applied to current task status';
-export const NO_RESPONSE_CONTRACTOR_ERROR =
-  'Contractor has not responded to task';
-export const CONTRACTOR_NOT_FREE_ERROR = 'Contractor has another task to do';
 
 export const TASK_TITLE_MIN_LENGTH = 20;
 export const TASK_TITLE_MAX_LENGTH = 50;
