+++ taskforce/apps/tasks/src/app/task/task.controller.ts
@@ -24,7 +24,6 @@ import {
   RolesGuard,
 } from '@taskforce/core';
 import { UserRole, UserRoles } from '@taskforce/shared-types';
-import { ChangeTaskStatusDto } from './dto/change-task-status.dto';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
 import { TaskRdo } from './rdo/task.rdo';
@@ -138,32 +137,6 @@ export class TaskController {
           TaskRdo,
           await this.taskService.findByContractor(userId, query)
         );
-  }
-
-  @Patch('task/:id/status')
-  @ApiResponse({
-    type: TaskRdo,
-    status: HttpStatus.OK,
-    description: 'Task status successfully updated',
-  })
-  @ApiResponse({
-    status: HttpStatus.NOT_FOUND,
-    description: 'Task with give id not found',
-  })
-  @UseGuards(JwtAuthGuard)
-  async changeTaskStatus(
-    @GetUser('id') userId: string,
-    @GetUser('role') userRole: UserRole,
-    @Param('id') taskId: number,
-    @Body() dto: ChangeTaskStatusDto
-  ) {
-    const updatedTask = await this.taskService.changeTaskStatus(
-      userId,
-      userRole,
-      taskId,
-      dto
-    );
-    return fillObject(TaskRdo, updatedTask);
   }
 
   @Post('task/:id/image')
