+++ taskforce/apps/tasks/src/app/task/task.module.ts
@@ -9,12 +9,10 @@ import { ConfigService } from '@nestjs/config';
 import { getRabbitMqOptions } from '@taskforce/config';
 import { MulterModule } from '@nestjs/platform-express';
 import { getMulterOptions } from '@taskforce/config';
-import { ResponseModule } from '../response/response.module';
 
 @Module({
   imports: [
     CommentModule,
-    ResponseModule,
     ClientsModule.registerAsync([
       {
         name: RABBITMQ_SERVICE,
@@ -29,6 +27,6 @@ import { ResponseModule } from '../response/response.module';
   ],
   controllers: [TaskController],
   providers: [TaskRepository, TaskService, ConfigService],
-  exports: [TaskService],
+  exports: [TaskService, TaskRepository],
 })
 export class TaskModule {}
