+++ taskforce/apps/tasks/src/app/tasks.http
@@ -1,7 +1,7 @@
 # Create a new task
 POST http://localhost:3335/api/tasks/task HTTP/1.1
 Content-Type: application/json
-Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2M2I3YzdjNTc3OGMzNTJhOWYyZTM2Y2IiLCJlbWFpbCI6InVzZXIyQHRlc3QuY29tIiwicm9sZSI6IkN1c3RvbWVyIiwibmFtZSI6Iktla3MiLCJpYXQiOjE2NzI5OTc3NzksImV4cCI6MTY3MzAwMTM3OX0.6a93_zkNV1hQiSJQk8930NBJ1GzpRzH_aTMp6ZfDhkU
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2M2I3YzdjNTc3OGMzNTJhOWYyZTM2Y2IiLCJlbWFpbCI6InVzZXIyQHRlc3QuY29tIiwicm9sZSI6IkN1c3RvbWVyIiwibmFtZSI6Iktla3MiLCJpYXQiOjE2NzMwMTExNTksImV4cCI6MTY3MzAxNDc1OX0.zFHZb-1S6P64BMOhMAf5BmVCqUJ2goYpX8Aww3JqYC0
 
 {
   "title": "Task's title veeeeeeeeeeeeeery long",
@@ -62,13 +62,13 @@ Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2M2I3YzdhM
 ###
 
 # Change task's status
-PATCH http://localhost:3335/api/tasks/task/9/status HTTP/1.1
+PATCH http://localhost:3335/api/tasks/task/11/status HTTP/1.1
 Content-Type: application/json
-Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2M2I3YzdjNTc3OGMzNTJhOWYyZTM2Y2IiLCJlbWFpbCI6InVzZXIyQHRlc3QuY29tIiwicm9sZSI6IkN1c3RvbWVyIiwibmFtZSI6Iktla3MiLCJpYXQiOjE2NzI5OTc3NzksImV4cCI6MTY3MzAwMTM3OX0.6a93_zkNV1hQiSJQk8930NBJ1GzpRzH_aTMp6ZfDhkU
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2M2I3YzdjNTc3OGMzNTJhOWYyZTM2Y2IiLCJlbWFpbCI6InVzZXIyQHRlc3QuY29tIiwicm9sZSI6IkN1c3RvbWVyIiwibmFtZSI6Iktla3MiLCJpYXQiOjE2NzMwMTExNTksImV4cCI6MTY3MzAxNDc1OX0.zFHZb-1S6P64BMOhMAf5BmVCqUJ2goYpX8Aww3JqYC0
 
 {
   "command": "start",
-  "contractor": "63b325e47cd9f2573b2cf15d"
+  "contractor": "63b8206a90d4c007a3dab6ab"
 }
 
 ###
@@ -107,9 +107,8 @@ GET http://localhost:3335/api/tasks/category HTTP/1.1
 ###
 
 # Create new contractor's response
-POST http://localhost:3335/api/tasks/task/8/response HTTP/1.1
-Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2M2I3YzdhMTc3OGMzNTJhOWYyZTM2YzciLCJlbWFpbCI6InVzZXIxQHRlc3QuY29tIiwicm9sZSI6IkNvbnRyYWN0b3IiLCJuYW1lIjoiS2VrcyIsImlhdCI6MTY3Mjk4OTg5MCwiZXhwIjoxNjcyOTkzNDkwfQ.As2QjuaJCrlSP8AmnZNLbVYo_9IdHoEJZNRRc8mAFtM
-
+POST http://localhost:3335/api/tasks/task/11/response HTTP/1.1
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2M2I4MjA2YTkwZDRjMDA3YTNkYWI2YWIiLCJlbWFpbCI6InVzZXIzQHRlc3QuY29tIiwicm9sZSI6IkNvbnRyYWN0b3IiLCJuYW1lIjoiS2VrcyIsImlhdCI6MTY3MzAxMTM0NiwiZXhwIjoxNjczMDE0OTQ2fQ.gvsv8zUIi3KaiSNmWI1fvgMOMJ69kdoi5hETzJ8YXqI
 ###
 
 # Get all responses by task
